% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata.R
\name{import_metadata}
\alias{import_metadata}
\title{Import metadata}
\usage{
import_metadata(sourcedir, metadata)
}
\arguments{
\item{sourcedir}{Path to directory containing metadata in .csv format}

\item{metadata}{Name of .csv file containing metadata}
}
\value{
a dataframe containing metadata of the experiment
}
\description{
Imports metadata in .csv format, needs a minimum of 3 columns with the following headings:
}
\details{
1: ID: Sample names corresponding to the sample names in the count matrix.

2: Genotype
May be left blank. Genotype of sample, e.g. WT or SOD1

3: Treatment
May be left blank. Treatment of sample, e.g. drug or placebo
}
\examples{
metadata <- import_metadata(sourcedir="C:/Documents/",metadata="metadata.csv")

}
\keyword{metadata}
