% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seqdata.R
\name{import_seqdata}
\alias{import_seqdata}
\title{Import seqdata}
\usage{
import_seqdata(sourcedir, countMatrix)
}
\arguments{
\item{countMatrix}{Path to directory containing your count matrix in .csv format}

\item{seqdata}{Name of your .csv file}
}
\value{
a dataframe of count data from a .csv file
}
\description{
Imports RNA-seq count matrix in .csv form
NB! Make sure the column with the gene identifiers (e.g. ENSEMBL ID, EntrezID,
NCBI accession number) is named!
}
\examples{
seqdata <- import_seqdata(sourcedir="C:/Documents/",countMatrix="countMatrix.csv")

 @export
}
\keyword{Count}
\keyword{matrix}
