% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_countMatrix.R
\name{process_countMatrix}
\alias{process_countMatrix}
\title{Complete pipeline for processing RNA-seq data in count matrix}
\usage{
process_countMatrix(
  sourcedir,
  countMatrix,
  metadata,
  rownameColumn,
  threshold,
  contrasts = TRUE,
  figures = TRUE,
  libsize = FALSE,
  destdir
)
}
\arguments{
\item{sourcedir}{Path to directory containing count matrix and metadata in .csv format}

\item{countMatrix}{Name of .csv file containing count matrix
NB! In the count matrix file, column of gene identifier (e.g. ENSEMBL ID, EntrezID,
NCBI accession number) needs to be named.}

\item{metadata}{Name of .csv contianing metadata.
NB! Metadata file requires a minimum of the following three columns:
1: ID: Sample names corresponding to the sample names in the count matrix.
2: Genotype: May be left blank. Genotype of sample, e.g. WT or SOD1
3: Treatment: May be left blank. Treatment of sample, e.g. drug or placebo}

\item{rownameColumn}{Name of column in count matrix file contianing gene identifiers}

\item{threshold}{Positive real number; threshold value of minimum counts per million
included in analysis}

\item{contrasts}{Logical. Determines whether logfold changes between two groups are calculated
with empirical Bayes shrinkage of variances and estimation of moderated t-statistic and p-values}

\item{figures}{Logical. Determines whether .png files of plots are saved}

\item{libsize}{Logical. Determines whether a .csv file of the library sizes for each sample in the
data set is saved}

\item{destdir}{Path to desired output directory}
}
\value{
A .csv file of normalised Log2 counts per million, plus optional contrast .csv files and figures
}
\description{
Differential expression analysis with limma-voom of RNA-seq data in count matrix
Based on pipeline by Phipson et al., 2016 (RNA-seq analysis in R)
}
\examples{
process_countMatrix(sourcedir="C:/Documents/",countMatrix="countMatrix.csv",
metadata="metadata.csv",rownameColumn="ENSEMBL",threshold=0.5,contrasts=TRUE,
figures=TRUE,libsize=FALSE,destdir="C:/Documents/Analysed/")

}
\references{
Phipson, B., Trigos, A., Ritchie, M, Doyle, M, Dashnow, H, Law, C. (2016).
RNA-seq analysis in R. \url{https://combine-australia.github.io/RNAseq-R/06-rnaseq-day1.html}

Ritchie, M.E., Phipson, B., Wu, D., Hu, Y., Law, C.W., Shi, W., and Smyth, G.K. (2015).
limma powers differential expression analyses for RNA-sequencing and microarray studies.
Nucleic Acids Research 43(7), e47.

Su, S., Law, C. W., Ah-Cann, C., Asselin-Labat, M. L., Blewitt, M. E., & Ritchie, M. E. (2017).
Glimma: interactive graphics for gene expression analysis. Bioinformatics, 33(13), 2050-2052.

Gregory R. Warnes, Ben Bolker, Lodewijk Bonebakker, Robert Gentleman, Wolfgang Huber, Andy Liaw,
Thomas Lumley, Martin Maechler, Arni Magnusson, Steffen Moeller, Marc Schwartz and Bill Venables (2020).
gplots: Various R Programming Tools for Plotting Data. R package version 3.0.3.
https://CRAN.R-project.org/package=gplots

Erich Neuwirth (2014). RColorBrewer: ColorBrewer Palettes. R package version 1.1-2.
https://CRAN.R-project.org/package=RColorBrewer

Ron Ammar (2019). randomcoloR: Generate Attractive Random Colors. R package version 1.1.0.1.
https://CRAN.R-project.org/package=randomcoloR

Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2020).
dplyr: A Grammar of Data Manipulation. R package version 1.0.0. https://CRAN.R-project.org/package=dplyr

Robinson MD, McCarthy DJ and Smyth GK (2010).
edgeR: a Bioconductor package for differential expression analysis of digital gene expression data.
Bioinformatics 26, 139-140

McCarthy DJ, Chen Y and Smyth GK (2012). Differential expression analysis of
multifactor RNA-Seq experiments with respect to biological variation.
Nucleic Acids Research 40, 4288-4297
}
\keyword{}
